👺 Havoc: Lock Ordering Deadlock Simulation & Model in ShardCoordinator
Replaced a fake inline "chaos" test (`test_havoc_deadlock`) in `src/storage/sharding/coordinator.rs` with a rigorous Loom concurrency model in `tests/havoc_loom_shard_coordinator.rs`. The previous inline test used `std::thread` and `std::sync::RwLock` on Linux, which defaults to reader-preference, causing the test to falsely assert that the missing `drop(connections)` before `active_transactions.write()` did not deadlock.

The faux test gave false confidence by asserting `rx.recv_timeout(Duration::from_secs(2)).is_ok()` (i.e. NO deadlock). To truly verify that the lock ordering invariant (drop `connections` before acquiring `active_transactions.write()`) holds under all thread interleavings (including writer starvation scenarios), a `loom` model must be used.
